ПОЧИНИЛ!!!

+++ Assets/Scripts/Model/Weapon.cs
@@ -9,9 +9,9 @@ public class Weapon
 {
     public int idweapon { get; set; }
     public int level { get; set; }
-    public double damage { get; set; }
-    public double distance { get; set; }
+    public float damage { get; set; }
+    public float distance { get; set; }
     public int attacktype { get; set; }
     public int projectile { get; set; }
-    public double attackspeed { get; set; }
+    public float attackspeed { get; set; }
 }
